refactor: split monolithic database.rs into a module directory
